+++ model/consume_status.go
@@ -1,0 +1,6 @@
+package model
+
+type ConsumeStatus struct {
+	GroupID  string
+	LagCount int
+}