+++ service/init_service.go
@@ -5,7 +5,7 @@ import (
 	"gitlab.paradise-soft.com.tw/glob/dispatcher/model"
 )
 
-var _ver = "v1.12.0"
+var _ver = "v1.13.beta"
 
 func Init(brokers []string, opts []model.Option) error {
 	dis := model.MakeDispatcher(opts)